+++ src/EntityGraphQL/Schema/MutationType.cs
@@ -134,7 +134,7 @@ namespace EntityGraphQL.Schema
         }
 
         /// <summary>
-        /// Used at runtime below
+        /// Used at runtime below!!
         /// </summary>
         /// <param name="input"></param>
         /// <typeparam name="T"></typeparam>